requirements.txt
apturl==0.5.2
asgiref==3.8.1
bcrypt==3.2.0
black==25.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cssbeautifier==1.15.4
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djlint==1.36.4
duplicity==0.8.21
EditorConfig==0.17.1
Faker==37.3.0
fasteners==0.14.1
flake8==4.0.1
future==0.18.2
git-filter-repo==2.34.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jsbeautifier==1.15.4
json5==0.12.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
mutagen==1.45.1
mypy_extensions==1.1.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==25.0
paramiko==2.9.3
pathspec==0.12.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.8
protobuf==3.12.4
psutil==5.9.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.8.0
pycryptodomex==3.11.0
pycups==2.0.1
pyflakes==2.4.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0.2
regex==2024.11.6
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.1
systemd-python==234
terminator==2.1.1
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
websockets==9.1
xdg==5
xkit==0.0.0
yt-dlp==2025.3.31
zipp==1.0.0
